Templates: Rewrite template lookup to optimise.
This commit completely restructures the template lookup algorithm for better
performance. In `lookup`, we no longer recursively call a local function
(`find`), but instead consume the path atoms in a loop. The template dictionary
has also been significantly simplified. Keys are now strings or Guards. Guards
are simpler than the old Keys and the restructure obviates the need for the
expensive dictionary specialisation that was performed previously.

On a simpler example, this change makes very little difference. However, on a
larger spec that involves many template lookups these costs can dominate. As an
example, for the current VM project, this shaves 26 seconds off compilation
time. This is more or less the entirety of the template lookup cost. After this
change the most expensive operation in a profile of the VM compilation is the
actual template rendering.