libsel4camkes: Bump up the allocated size to meet the alignment constraint.
CAmkES DMA allocator divides a large chunk DMA memory into small pieces for
small allocation and stores the bookkeeping info at the beginning of each
piece. The current implementation only works if the address of each piece is
aligned to the bookkeeping structure for efficiency. If the user allocates an
unaligned size, the allocator would fail. So we automatic bump up the size
instead of expose the constraint to the user.